INS-213: move storage test utils function in separte file.
